ビルド時に表示される場合があるwarningを修正
In file included from src\robot\main.cpp:5:0:
.pio\libdeps\robot\liboshima\src/components/ImReceiver.h: In instantiation of 'bool ImReceiver::receive(T&) [with T = Controller]':
src\robot\main.cpp:24:30:   required from here
.pio\libdeps\robot\liboshima\src/components/ImReceiver.h:42:30: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
       if (recvedStr.length() > recvedStrLen) {
           ~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~
.pio\libdeps\robot\liboshima\src/components/ImReceiver.h:63:29: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     if (recvedData.length() != hexStrLen + commaCount) {
         ~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~

+++ src/components/ImReceiver.h
@@ -39,7 +39,7 @@ public:
       else
         recvedStr += recvedChar;
 
-      if (recvedStr.length() > recvedStrLen) {
+      if (recvedStr.length() > static_cast<size_t>(recvedStrLen)) {
         DebugLogger::println("ImReceiver", "receive",
                              "Received string is too long");
         return false;
@@ -60,7 +60,7 @@ public:
     String recvedData = recvedStr.substring(colonIndex + 1);
 
     // データが長さが適切でない場合は false を返す
-    if (recvedData.length() != hexStrLen + commaCount) {
+    if (recvedData.length() != static_cast<size_t>(hexStrLen + commaCount)) {
       DebugLogger::println("ImReceiver", "receive", "Data length is invalid");
       return false;
     }
